# api/openapi.yaml
openapi: 3.1.0
info:
  title: Crabit API
  version: 0.0.1
  description: |
    Owner-facing Card Balance Account and Wish operations plus academy-facing
    Shared Card reads. Resource ownership and current social visibility are
    hidden with resource-specific 404 responses. All timestamps are RFC 3339
    UTC Z values, all KRW values are bounded integer won, and all error bodies
    use the common ErrorEnvelope.
servers:
  - url: /
...
paths:
  /v1/me/card-balance-accounts:
    get:
      tags: [Card Balance Accounts]
      operationId: listMyCardBalanceAccounts
      summary: List the authenticated student's Card Balance Accounts
      description: UNKNOWN balances remain null rather than being fabricated as zero. Each account also reports whether an account-scoped Balance Adjustment Case is currently OPEN.
      security: [{SeedBearer: []}]
      responses:
        '200':
          description: Visible Card Balance Accounts.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CardBalanceAccountPage'}
              examples:
                unknown-balance: {$ref: '#/components/examples/UnknownBalancePage'}
        '401': {$ref: '#/components/responses/AuthRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/card-balance-accounts/{cardBalanceAccountId}/balance-refreshes:
    parameters: [{$ref: '#/components/parameters/CardBalanceAccountId'}]
    post:
      tags: [Card Balance Accounts]
      operationId: refreshCardBalance
      summary: Refresh the current card balance
      description: Bodyless USER_REQUESTED lookup. This operation is deliberately not idempotency-keyed and remains allowed while a Balance Adjustment Case is OPEN; the response reports the resulting current adjustment flag.
      security: [{SeedBearer: []}]
      responses:
        '200':
          description: A successful current balance observation.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BalanceRefreshResult'}
        '401': {$ref: '#/components/responses/AuthRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/CardBalanceAccountNotFound'}
        '503': {$ref: '#/components/responses/BalanceSyncFailed'}
...
components:
  securitySchemes:
    SeedBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-seed-token
      description: Opaque deterministic principal token. A known Seed token identifies either a student or an authenticated non-student staff principal. Token issuance and refresh are outside this contract.
  parameters:
    CardBalanceAccountId:
      name: cardBalanceAccountId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Uuid'}
...
  responses:
...
    AuthRequired:
      description: AUTH_REQUIRED — missing or invalid bearer token.
      x-error-codes: [AUTH_REQUIRED]
      headers:
        WWW-Authenticate:
          required: true
          schema: {type: string, const: Bearer}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    Forbidden:
      description: FORBIDDEN — the authenticated principal is not a student.
      x-error-codes: [FORBIDDEN]
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    CardBalanceAccountNotFound:
      description: CARD_BALANCE_ACCOUNT_NOT_FOUND — absent or non-owned account is hidden.
      x-error-codes: [CARD_BALANCE_ACCOUNT_NOT_FOUND]
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
    BalanceSyncFailed:
      description: BALANCE_SYNC_FAILED — retryable external balance query failure; the failed observation is persisted without mutating the Wish.
      x-error-codes: [BALANCE_SYNC_FAILED]
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
          examples:
            balance-sync-failed: {$ref: '#/components/examples/BalanceSyncFailed'}
  schemas:
    Uuid:
      type: string
      format: uuid
    UtcInstant:
      type: string
      format: date-time
      pattern: 'Z$'
      description: RFC 3339 UTC instant whose wire representation ends in Z.
...
    KrwSigned:
      type: integer
      format: int64
      minimum: -9007199254740991
      maximum: 9007199254740991
      description: Fractional KRW is never accepted.
    KrwNonNegative:
      type: integer
      format: int64
      minimum: 0
      maximum: 9007199254740991
...
    ErrorCode:
      type: string
      enum:
        - MALFORMED_REQUEST
        - EXPECTED_VERSION_REQUIRED
        - IDEMPOTENCY_KEY_REQUIRED
        - AUTH_REQUIRED
        - FORBIDDEN
        - CARD_BALANCE_ACCOUNT_NOT_FOUND
        - WISH_NOT_FOUND
        - ACADEMY_NOT_FOUND
        - SHARED_CARD_NOT_FOUND
        - VERSION_CONFLICT
        - INVALID_STATE_TRANSITION
        - BALANCE_MISMATCH_LOCKED
        - INSUFFICIENT_AVAILABLE_BALANCE
        - INSUFFICIENT_WISH_AMOUNT
        - TARGET_AMOUNT_EXCEEDED
        - CROSS_ACCOUNT_TRANSFER_FORBIDDEN
        - IDEMPOTENCY_KEY_REUSED
        - UNSUPPORTED_MEDIA_TYPE
        - INVALID_AMOUNT
        - INVALID_PURPOSE
        - INVALID_VERSION
        - BALANCE_SYNC_FAILED
    FieldError:
      type: object
      additionalProperties: false
      required: [field, message]
      properties:
        field: {type: string, minLength: 1, description: "Name of the invalid request field, parameter, or header associated with this validation failure."}
        message: {type: string, minLength: 1, description: "Human-readable explanation of the field-specific failure."}
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          description: "Structured error payload shared by every declared non-success JSON response."
          type: object
          additionalProperties: false
          required: [code, message, retryable, traceId, fieldErrors, details]
          allOf:
            - if:
                properties:
                  code: {const: BALANCE_SYNC_FAILED}
              then:
                properties:
                  retryable: {const: true}
              else:
                properties:
                  retryable: {const: false}
          properties:
            code: {$ref: '#/components/schemas/ErrorCode', description: "Stable machine-readable ErrorCode; clients should branch on this value rather than message text."}
            message: {type: string, minLength: 1, description: "Human-readable explanation of this occurrence; it is not the stable machine decision key."}
            retryable:
              type: boolean
              description: True only for BALANCE_SYNC_FAILED; false for every defined client, authorization, not-found, validation, and state-conflict error.
            traceId: {type: string, minLength: 1, description: "Opaque server correlation identifier for diagnostics and support; it has no domain meaning."}
            fieldErrors:
              description: "Field-specific validation failures; empty when the error is not attributable to individual request fields."
              type: array
              items: {$ref: '#/components/schemas/FieldError'}
            details:
              description: "Extensible code-specific metadata object; empty when no details apply, and clients must ignore unrecognized keys."
              type: object
              additionalProperties: true
    CardBalanceAccount:
      oneOf:
        - $ref: '#/components/schemas/UnknownCardBalanceAccount'
        - $ref: '#/components/schemas/KnownCardBalanceAccount'
      discriminator:
        propertyName: balanceKnowledge
        mapping:
          UNKNOWN: '#/components/schemas/UnknownCardBalanceAccount'
          KNOWN: '#/components/schemas/KnownCardBalanceAccount'
    UnknownCardBalanceAccount:
      type: object
      additionalProperties: false
      required: [cardBalanceAccountId, academyId, balanceKnowledge, actualCardBalance, ledgerAvailableBalance, displayAvailableBalance, unresolvedShortage, balanceAdjustmentInProgress, lastRefreshStatus, lastRefreshedAt]
      properties:
        cardBalanceAccountId: {$ref: '#/components/schemas/Uuid', description: "Stable UUID of the student's Card Balance Account."}
        academyId: {$ref: '#/components/schemas/Uuid', description: "UUID of the academy to which this Card Balance Account belongs."}
        balanceKnowledge: {type: string, const: UNKNOWN, description: "UNKNOWN means no successful external balance observation exists; null balance values must never be interpreted as zero."}
        actualCardBalance: {type: 'null', description: "Always null because no successful external balance observation exists; null means unknown, not zero KRW."}
        ledgerAvailableBalance: {type: 'null', description: "Always null until actual card balance is known; the service must not calculate it from a fabricated zero."}
        displayAvailableBalance: {type: 'null', description: "Always null until actual card balance is known; the UI must not display it as zero KRW."}
        unresolvedShortage: {type: 'null', description: "Always null until actual card balance is known; unknown shortage is not the same as no shortage."}
        balanceAdjustmentInProgress: {type: boolean, const: false, description: "Always false because an OPEN Balance Adjustment Case requires a successful balance observation, which an UNKNOWN account does not have."}
        lastRefreshStatus:
          description: "FAILED when the latest lookup attempt failed before any successful observation; null when no lookup attempt has been recorded."
          type: [string, 'null']
          enum: [FAILED, null]
        lastRefreshedAt:
          description: "Always null because no successful observation exists; failed-attempt time is represented in observation history instead."
          type: ['null']
    KnownCardBalanceAccount:
      type: object
      additionalProperties: false
      required: [cardBalanceAccountId, academyId, balanceKnowledge, actualCardBalance, ledgerAvailableBalance, displayAvailableBalance, unresolvedShortage, balanceAdjustmentInProgress, lastRefreshStatus, lastRefreshedAt]
      properties:
        cardBalanceAccountId: {$ref: '#/components/schemas/Uuid', description: "Stable UUID of the student's Card Balance Account."}
        academyId: {$ref: '#/components/schemas/Uuid', description: "UUID of the academy to which this Card Balance Account belongs."}
        balanceKnowledge: {type: string, const: KNOWN, description: "KNOWN means at least one successful external balance observation supplies the returned balance values."}
        actualCardBalance: {$ref: '#/components/schemas/KrwNonNegative', description: "Non-negative integer KRW observed by the most recent successful external card-balance lookup."}
        ledgerAvailableBalance: {$ref: '#/components/schemas/KrwSigned', description: "Signed integer KRW equal to actualCardBalance minus the total held by active Wishes; a negative value is preserved."}
        displayAvailableBalance: {$ref: '#/components/schemas/KrwNonNegative', description: "Non-negative integer KRW equal to max(0, ledgerAvailableBalance) for display."}
        unresolvedShortage: {$ref: '#/components/schemas/KrwNonNegative', description: "Non-negative integer KRW equal to max(0, -ledgerAvailableBalance); zero means no unresolved shortage."}
        balanceAdjustmentInProgress: {type: boolean, description: "True iff an account-scoped Balance Adjustment Case is OPEN at response read time; false when no OPEN case exists, including RESOLVED-only history. Derived rather than persisted on the account, this flag and the balance fields come from one consistent account projection; a later failed lookup retains the latest successful amounts while the flag reflects the current case."}
        lastRefreshStatus: {type: string, enum: [SUCCESS, FAILED], description: "Outcome of the most recent lookup attempt; FAILED may coexist with amounts preserved from an earlier successful observation."}
        lastRefreshedAt: {$ref: '#/components/schemas/UtcInstant', description: "RFC 3339 UTC Z instant of the successful observation backing the returned amounts; retained when a later lookup fails."}
    CardBalanceAccountPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          description: "Card Balance Accounts visible to the authenticated student in this page."
          type: array
          items: {$ref: '#/components/schemas/CardBalanceAccount'}
        nextCursor:
          description: "Opaque cursor for the next account page; null when no further page exists."
          type: [string, 'null']
          minLength: 1
    BalanceRefreshResult:
      type: object
      additionalProperties: false
      required: [observationId, lookupMethod, observedAt, account]
      properties:
        observationId: {$ref: '#/components/schemas/Uuid', description: "UUID of the newly persisted successful balance observation."}
        lookupMethod:
          description: "Always USER_REQUESTED because this public bodyless operation performs a user-requested lookup and does not accept a client-selected method."
          type: string
          const: USER_REQUESTED
        observedAt: {$ref: '#/components/schemas/UtcInstant', description: "RFC 3339 UTC Z instant at which this external balance lookup attempt was made."}
        account: {$ref: '#/components/schemas/KnownCardBalanceAccount', description: "Updated KNOWN Card Balance Account snapshot derived from this successful observation."}
...
  examples:
    UnknownBalancePage:
      summary: unknown-balance
      x-schema-ref: '#/components/schemas/CardBalanceAccountPage'
      value:
        items:
          - cardBalanceAccountId: 09551375-f69a-4e69-8587-a0d590cbf092
            academyId: a12c54ad-c8a2-4102-a03b-28fe4a0daaa7
            balanceKnowledge: UNKNOWN
            actualCardBalance: null
            ledgerAvailableBalance: null
            displayAvailableBalance: null
            unresolvedShortage: null
            balanceAdjustmentInProgress: false
            lastRefreshStatus: null
            lastRefreshedAt: null
        nextCursor: null
...
    BalanceSyncFailed:
      summary: balance-sync-failed
      x-schema-ref: '#/components/schemas/ErrorEnvelope'
      value:
        error:
          code: BALANCE_SYNC_FAILED
          message: Card balance could not be refreshed.
          retryable: true
          traceId: tr_01J5G1W1M9QH7QVCATBMRXJRGC
          fieldErrors: []
          details: {}
...